# apps/api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  inecgrabber.ResolveUnmatchedRequest:
    properties:
      notes:
        type: string
      polling_unit_id:
        type: integer
      status:
        description: resolved_mapped, resolved_created_pu, rejected
        type: string
    type: object
...
host: localhost:4100
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for the Free9ja platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Free9ja API
  version: "1.0"
paths:
...
  /admin/inec-result-grabber-logs:
    get:
      consumes:
      - application/json
      description: Retrieves paginated execution logs across grabbers
      parameters:
      - description: Filter by grabber ID
        in: query
        name: grabber_id
        type: integer
      - description: Limit results (default 20)
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Logs retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all INEC result grabber logs
      tags:
      - INEC Result Grabber
  /admin/inec-result-grabbers:
    get:
      consumes:
      - application/json
      description: Retrieves paginated INEC result grabbers with election details
        and metrics
      parameters:
      - description: Limit results (default 20)
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: INEC result grabbers retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List INEC result grabbers
      tags:
      - INEC Result Grabber
  /admin/inec-result-grabbers/{id}/logs:
    get:
      consumes:
      - application/json
      description: Retrieves paginated execution logs for a specific INEC result grabber
      parameters:
      - description: INEC Result Grabber ID
        in: path
        name: id
        required: true
        type: integer
      - description: Limit results (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset results (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Logs retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid grabber ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List INEC result grabber sync logs
      tags:
      - INEC Result Grabber
  /admin/inec-result-grabbers/{id}/sync:
    post:
      consumes:
      - application/json
      description: Triggers a manual sync run for the specified INEC result grabber
      parameters:
      - description: INEC Result Grabber ID
        in: path
        name: id
        required: true
        type: integer
      - description: Upload result sheet images to R2 (default from config)
        in: query
        name: upload_to_r2
        type: boolean
      - description: Extract candidate results via AI (default from config)
        in: query
        name: ai_extract
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Sync completed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request params
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Manually trigger INEC result grabber sync
      tags:
      - INEC Result Grabber
  /admin/inec-result-grabbers/{id}/toggle-pause:
    post:
      parameters:
      - description: INEC Result Grabber ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Toggle pause/resume status of INEC result grabber
      tags:
      - INEC Result Grabber
...
  /admin/unmatched-polling-unit-results:
    get:
      consumes:
      - application/json
      description: Retrieves paginated queue of results submitted without a matching
        polling unit in the system
      parameters:
      - description: Filter by election ID
        in: query
        name: election_id
        type: integer
      - description: Filter by status (pending, resolved_mapped, resolved_created_pu,
          rejected)
        in: query
        name: resolution_status
        type: string
      - description: Limit results (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset results (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Unmatched results retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List unmatched polling unit results
      tags:
      - INEC Result Grabber
  /admin/unmatched-polling-unit-results/{id}/resolve:
    post:
      consumes:
      - application/json
      description: Resolves an unmatched result by mapping to an existing/new polling
        unit or rejecting it
      parameters:
      - description: Unmatched Result ID
        in: path
        name: id
        required: true
        type: integer
      - description: Resolution payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/inecgrabber.ResolveUnmatchedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Unmatched result resolved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload or ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Resolve unmatched polling unit result
      tags:
      - INEC Result Grabber
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your access token.
    in: header
    name: Authorization
    type: apiKey
  RefreshToken:
    description: Type your refresh token.
    in: header
    name: RefreshToken
    type: apiKey
swagger: "2.0"